Rename ui/src/graphql/schema.ts to schema.d.ts. Other misc tweaks and cleanup in incoming files
