+++ ui/src/pages/CollectionsManagement.tsx
@@ -27,7 +27,7 @@ import { Link } from 'src/components/Link.tsx';
 
 import { LoadingBlock } from 'src/components/Loading.tsx';
 import { GET_COLLECTION_DETAILS } from 'src/graphql/queries.ts';
-import type { ColumnDef } from 'src/graphql/schema.ts';
+import type { ColumnDef } from 'src/graphql/schema.d.ts';
 
 const ErrorDisplay = function ({
   title,
