+++ ui/src/pages/ColumnManagement.tsx
@@ -13,7 +13,7 @@ import ColumnManagementForm from 'src/components/ColumnManagementForm.tsx';
 import { Link } from 'src/components/Link.tsx';
 import { LoadingBlock } from 'src/components/Loading.tsx';
 import { GET_COLUMN_DETAILS } from 'src/graphql/queries.ts';
-import type { User } from 'src/graphql/schema.ts';
+import type { User } from 'src/graphql/schema.d.ts';
 
 const ErrorDisplay = function ({
   title,
