+++ ui/src/pages/Home.tsx
@@ -35,7 +35,7 @@ import { useSession } from 'src/components/auth/session.tsx';
 import { Link } from 'src/components/Link.tsx';
 import { LoadingBlock } from 'src/components/Loading.tsx';
 import { GET_HOME_INFO } from 'src/graphql/queries.ts';
-import type { CollectionInfo } from 'src/graphql/schema.ts';
+import type { CollectionInfo } from 'src/graphql/schema.d.ts';
 
 const CollectionTable = ({
   tableCaption,
@@ -145,6 +145,7 @@ const HomeDynamic = memo(function HomeDynamic({
 
   const { loading, error, data, refetch } = useQuery(GET_HOME_INFO, {
     variables: { lang: locale },
+    fetchPolicy: 'no-cache',
   });
 
   if (!loading && data.session === null) {
